+++ DeliveryCompany.DataLayer/Models/Rating.cs
@@ -7,6 +7,6 @@ namespace DeliveryCompany.DataLayer.Models
         public int Id { get; set; }
         public int UserId { get; set; }
         public DateTime DateTime { get; set; }
-        public int CouriersRating { get; set; }
+        public double CouriersRating { get; set; }
     }
 }
